+++ CHANGELOG.md
@@ -12,6 +12,9 @@ All notable changes to `laravel-event-sourcing` will be documented in this file:
 - you can now get the `uuid` of an aggregate root using the `uuid()` method
 - the `reset` method has been removed on projectors
 - the `fake` method on an aggregate root now accepts a uuid instead of an array of events
+- the `meta_data` property on `StoredEvent` is now an array or a string instead of `SchemalessAttributes`
+- apply methods on aggregates can now make use of method injection
+- pass metadata to serializer to allow events to be upgraded (#112)
 
 ## 3.1.4 - 2020-05-07